Fix database write issue by making workers run as root
The SQLite library initializes the database file as root because the NGINX
master-process runs as root. The connection is then put in shared memory,
and NGINX forks to create worker processes. But, workers fork as the user
`nobody' instead of root. The connection opened as root to a root-owned
database is being used by the `nobody' user causes the SQLite library to
crash and throw a SIGSEGV. We fix that by configuring NGINX to fork new
workers as root instead. There could be some better solution.