Fix Email IMAP search to use UIDs instead of message indexes
The default is to use a serial message index when doing searches. This is
bad, because if an Email was deleted from the inbox, then everything gets
messed up. Our code already assumes UIDs when fetching. To fix this, we use
a custom cURL `--request' parameter, with a custom IMAP command. UID SEARCH
causes search results to be returned by UID instead of index.